requirements.txt
django-bootstrap-v5
django-localflavor
